Made some casts explicit to avoid compiler warnings.  Removed code that wasn't being used.  Update for 1.7.3.

+++ Data Miners/APSL/SMCSensors.m
@@ -163,7 +163,7 @@ typedef enum {
 
 - (NSDictionary *)allValues
 {
-    int           totalKeys, i;
+    NSInteger     totalKeys, i;
     uint32_t      key;
   
 	NSMutableDictionary *dict = [NSMutableDictionary dictionary];
@@ -206,7 +206,7 @@ typedef enum {
     NSMutableSet *unknownTempKeys = [NSMutableSet set];
     NSMutableSet *knownTempKeys = [NSMutableSet set];
     
-	int smcKeyCount = [smc_ keyCount];
+	NSInteger smcKeyCount = [smc_ keyCount];
     NSMutableSet *availableKeys = [NSMutableSet setWithCapacity:smcKeyCount];
 
 	// traverse the available keys, prepare them for sorting
